batch working

+++ sdk/tables/aztables/zt_tableRecordedTests.go
@@ -138,24 +138,24 @@ func createSimpleEntities(count int, pk string) *[]map[string]interface{} {
 	return &result
 }
 
-func createComplexMapEntities(context *testContext, count int, pk string) *[]map[string]interface{}  {
+func createComplexMapEntities(context *testContext, count int, pk string) *[]map[string]interface{} {
 	result := make([]map[string]interface{}, count)
 
 	for i := 1; i <= count; i++ {
 		var e = map[string]interface{}{
-			PartitionKey: pk,
-			RowKey:       fmt.Sprint(i),
-			"StringProp": fmt.Sprintf("some string %d", i),
-			"IntProp":    i,
-			"BoolProp":   true,
-			"SomeBinaryProperty":    []byte("some bytes"),
-			"SomeDateProperty":      context.recording.Now(),
-			"SomeDoubleProperty0":   float64(1),
-			"SomeDoubleProperty1":   float64(1.2345),
-			"SomeGuidProperty":      context.recording.UUID(),
-			"SomeInt64Property":     math.MaxInt64,
-			"SomeIntProperty":       42,
-			"SomeStringProperty":    "some string",
+			PartitionKey:          pk,
+			RowKey:                fmt.Sprint(i),
+			"StringProp":          fmt.Sprintf("some string %d", i),
+			"IntProp":             i,
+			"BoolProp":            true,
+			"SomeBinaryProperty":  []byte("some bytes"),
+			"SomeDateProperty":    context.recording.Now(),
+			"SomeDoubleProperty0": float64(1),
+			"SomeDoubleProperty1": float64(1.2345),
+			"SomeGuidProperty":    context.recording.UUID(),
+			"SomeInt64Property":   (int64)(math.MaxInt64),
+			"SomeIntProperty":     42,
+			"SomeStringProperty":  "some string",
 		}
 		result[i-1] = e
 	}
